+++ storybook/themeManager.ts
@@ -1,3 +1,0 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
-import { defaultTheme, ThemeManager } from "@kaizen/design-tokens"
-export const themeManager = new ThemeManager(defaultTheme)