+++ draft-packages/title-block-zen/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.2.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block-zen@3.1.8...@kaizen/draft-title-block-zen@3.2.0) (2021-03-10)
+
+
+### Features
+
+* Include notification prop to display on Information Modal ([#1166](https://github.com/cultureamp/kaizen-design-system/issues/1166)) ([0295253](https://github.com/cultureamp/kaizen-design-system/commit/0295253ffe4ca2f60040f50866da19f426d6b7c0))
+
+
+
+
+
 ## [3.1.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block-zen@3.1.7...@kaizen/draft-title-block-zen@3.1.8) (2021-03-04)
 
 **Note:** Version bump only for package @kaizen/draft-title-block-zen