+++ packages/design-tokens/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/design-tokens@2.5.2...@kaizen/design-tokens@2.5.3) (2021-03-10)
+
+**Note:** Version bump only for package @kaizen/design-tokens
+
+
+
+
+
+## [2.5.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/design-tokens@2.5.1...@kaizen/design-tokens@2.5.2) (2021-03-09)
+
+**Note:** Version bump only for package @kaizen/design-tokens
+
+
+
+
+
 ## [2.5.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/design-tokens@2.5.0...@kaizen/design-tokens@2.5.1) (2021-02-25)